feat: move Text from display in canvas to above
