+++ client/src/components/UIManager.js
@@ -1,6 +1,7 @@
 import State from '../State.js';
 import { standingsTableRow } from '../utils/utils.js';
 import { Tournament } from './Tournament.js';
+import { GamePhases } from '../constants.js';
 
 export class UIManager {
 	constructor() {
@@ -34,16 +35,49 @@ export class UIManager {
 		this.content.fixtures = document.getElementById('fixtures');
 		this.content.standings = document.getElementById('standings');
 		this.content.standingsTable = document.getElementById('standingsTable');
+		this.content.gameInformation = document.getElementById('gameInformation');
 		this.update();
 	}
 
 	update() {
 		this.updateScoreCard()
+		this.updateGameInformation();
 		if (State.get('tournament')) {
 			this.updateMatchList();
 			this.updateTable();
 		} else {
 			this.hideTournamentInfo();
+		}
+	}
+
+	updateGameInformation() {
+		const currentMatch = State.get('currentMatch');
+
+		const gamePhase = State.get('gameData', 'phase');
+		switch (gamePhase) {
+			case GamePhases.WAITING_TO_START:
+				this.content.gameInformation.innerHTML = 'Press Enter to Start';
+				//this.drawTopText('Press Enter to Start');
+				break;
+			case GamePhases.COUNTDOWN:
+				this.content.gameInformation.innerHTML = Math.ceil(State.get("gameData", "countdown").toString());
+				//this.drawTopText(State.get("gameData", "countdown").toString());
+				break;
+			case GamePhases.RUNNING:
+				this.content.gameInformation.innerHTML = 'Running';
+				//this.drawTopText(State.get("gameData", "countdown").toString());
+				break;
+			case GamePhases.MATCH_ENDED:
+				this.content.gameInformation.innerHTML = 'Math Ended';
+				//const winner = currentMatch.players[0].score > currentMatch.players[1].score ? currentMatch.players[0] : currentMatch.players[1];
+				//this.drawTopText(`${winner.name} wins the match!`);
+				break;
+			case GamePhases.FINISHED:
+				this.content.gameInformation.innerHTML = 'Tournament Ended';
+				//this.drawTopText('Tournament Completed!');
+				//const tournamentWinner = currentMatch.players[0].score > currentMatch.players[1].score ? currentMatch.players[0] : currentMatch.players[1];
+				//this.drawBottomText(`${tournamentWinner.name} wins the tournament!`);
+				break;
 		}
 	}
