+++ client/templates/game.html
@@ -23,6 +23,10 @@
 			</div>
 		</div>
 	</div>
+	
+	<div>
+		<h5 class="text-center mb-3 f" id="gameInformation">Loading...</h5>
+	</div>
 
 	<!-- GAME CANVAS -->
 	<div class="row mb-3">
